+++ game/module_cargo.cpp
@@ -214,7 +214,7 @@ void mark::module::cargo::serialize(YAML::Emitter& out) const {
 	this->serialize_base(out);
 
 	out << Key << "contents" << Value << BeginSeq;
-	for (unsigned i = 0, size = m_modules.size(); i < size; i++) {
+	for (size_t i = 0, size = m_modules.size(); i < size; i++) {
 		const auto& module = m_modules[i];
 		if (module) {
 			out << BeginMap;
